+++ Core/Runtime/Extensions/RectTransform.cs
@@ -1,3 +1,4 @@
+using NorskaLib.Utilities;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -15,10 +16,12 @@ namespace NorskaLib.Extensions
             }
         }
 
-        [System.Obsolete("Use Transform.localScale = Vector3Utils.Uniform instead.")]
-        public static void SetScale(this RectTransform instance, float scale)
+        /// <returns> A vector from (0;0) to (1;1) where (0;0) matches top-left corner of the RectTransform.rect. </returns>
+        public static Vector2 NormalizedRectPostion(this RectTransform instance, Vector2 localPosition)
         {
-            instance.localScale = new Vector3(scale, scale, scale);
+            var localMinPosition = -1 * instance.pivot * instance.rect.size;
+            var localMaxPosition = (Vector2.one - instance.pivot) * instance.rect.size;
+            return Vector2Utils.InverseLerp(localMinPosition, localMaxPosition, localPosition);
         }
     }
 }